modules/url_resolver.py
# ==== Module Imports ====
import requests
from urllib.parse import urlparse, parse_qs
from bs4 import BeautifulSoup
import logging

# ==== Embedded URL Extraction ====
def extract_embedded_url(url):
    parsed = urlparse(url)
    query = parse_qs(parsed.query)
    return query.get('url', [None])[0]

# ==== Redirect Resolution via HEAD ====
def follow_redirects(url):
    try:
        response = requests.head(url, allow_redirects=True, timeout=5)
        if response.status_code in [301, 302] and 'Location' in response.headers:
            return response.headers['Location']
        return response.url
    except requests.RequestException as e:
        logging.warning(f"Redirect failed for {url}: {e}")
        return None

# ==== Canonical URL Extraction from HTML ====
def extract_canonical_from_html(url):
    try:
        headers = {'User-Agent': 'Mozilla/5.0'}
        resp = requests.get(url, headers=headers, timeout=8)
        soup = BeautifulSoup(resp.text, 'html.parser')
        canonical = soup.find('link', rel='canonical')
        if canonical and canonical.get('href'):
            return canonical['href']
    except Exception as e:
        logging.warning(f"Failed to extract canonical URL from {url}: {e}")
    return url

# ==== Final URL Resolver ====
def resolve_original_url(url):
    embedded = extract_embedded_url(url)
    if embedded:
        return embedded

    redirected = follow_redirects(url)
    if redirected and redirected != url:
        return redirected

    canonical = extract_canonical_from_html(url)
    return canonical or url
